requirements_txt.txt
# DINOv2 课堂演示项目依赖包

# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# Hugging Face Transformers
transformers>=4.30.0

# Web界面
gradio>=3.35.0

# 数据处理和可视化
numpy>=1.21.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
Pillow>=9.0.0

# 网络请求
requests>=2.28.0

# Jupyter支持 (可选)
jupyter>=1.0.0
ipywidgets>=8.0.0

# 性能优化 (可选)
accelerate>=0.20.0

# 图像处理增强 (可选)
opencv-python>=4.7.0